meme-generator-starter-code/src/requirements.txt
Flask==2.0.3
pandas==2.0.0
python-docx==1.1.2
requests==2.27.1
pillow==10.2.0